+++ app/(examples)/layout.tsx
@@ -1,5 +1,0 @@
-import type { PropsWithChildren } from 'react'
-
-export default function ExamplesLayout({ children }: PropsWithChildren) {
-  return <>{children}</>
-}